+++ CHANGELOG.md
@@ -49,6 +49,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Features
 
 - [2129](https://github.com/umee-network/umee/pull/2129) Emergency Group x/ugov proto.
+- [2146](https://github.com/umee-network/umee/pull/2146) Add store GetTimeMs and SetTimeMs.
 
 ### Improvements